initial migration of crop-related functionality out of field
